+++ include/envoy/ratelimit/ratelimit.h
@@ -2,8 +2,6 @@
 
 #include "envoy/common/optional.h"
 #include "envoy/common/pure.h"
-#include "envoy/runtime/runtime.h"
-#include "envoy/stats/stats.h"
 
 namespace RateLimit {
 
@@ -91,20 +89,5 @@ public:
 };
 
 typedef std::unique_ptr<ClientFactory> ClientFactoryPtr;
-
-/**
- * An interface for the rate limit filter configuration.
- */
-class FilterConfig {
-public:
-  virtual ~FilterConfig() {}
-
-  // TODO add comments
-  virtual const std::string& domain() const PURE;
-  virtual const std::string& localServiceCluster() const PURE;
-  virtual int64_t stage() const PURE;
-  virtual Runtime::Loader& runtime() PURE;
-  virtual Stats::Store& stats() PURE;
-};
 
 } // RateLimit
