+++ test/CMakeLists.txt
@@ -78,6 +78,7 @@ add_executable(envoy-test
   common/router/config_impl_test.cc
   common/router/retry_state_impl_test.cc
   common/router/router_test.cc
+  common/router/router_ratelimit_test.cc
   common/router/shadow_writer_impl_test.cc
   common/runtime/runtime_impl_test.cc
   common/runtime/uuid_util_test.cc
