+++ test/common/http/filter/ratelimit_test.cc
@@ -15,45 +15,11 @@ using testing::Invoke;
 using testing::NiceMock;
 using testing::Return;
 using testing::ReturnRef;
+using testing::SetArgReferee;
 using testing::WithArgs;
 
 namespace Http {
 namespace RateLimit {
-
-TEST(HttpRateLimitFilterBadConfigTest, BadType) {
-  std::string json = R"EOF(
-  {
-    "domain": "foo",
-    "actions": [
-      {"type": "foo"}
-    ]
-  }
-  )EOF";
-
-  Json::ObjectPtr config = Json::Factory::LoadFromString(json);
-  Stats::IsolatedStoreImpl stats_store;
-  NiceMock<Runtime::MockLoader> runtime;
-  EXPECT_THROW(FilterConfig(*config, "service_cluster", stats_store, runtime), EnvoyException);
-}
-
-TEST(HttpRateLimitFilterBadConfigTest, NoDescriptorKey) {
-  std::string json = R"EOF(
-  {
-    "domain": "foo",
-    "actions": [
-      {
-        "type": "request_headers",
-        "header_name" : "test"
-      }
-    ]
-  }
-  )EOF";
-
-  Json::ObjectPtr config = Json::Factory::LoadFromString(json);
-  Stats::IsolatedStoreImpl stats_store;
-  NiceMock<Runtime::MockLoader> runtime;
-  EXPECT_THROW(FilterConfig(*config, "service_cluster", stats_store, runtime), EnvoyException);
-}
 
 class HttpRateLimitFilterTest : public testing::Test {
 public:
@@ -73,36 +39,13 @@ public:
     client_ = new ::RateLimit::MockClient();
     filter_.reset(new Filter(config_, ::RateLimit::ClientPtr{client_}));
     filter_->setDecoderFilterCallbacks(filter_callbacks_);
+    rate_limit_policies_.clear();
+    rate_limit_policies_.emplace_back(rate_limit_policy_entry_);
   }
 
-  const std::string service_to_service_json = R"EOF(
+  const std::string filter_config = R"EOF(
     {
       "domain": "foo",
-      "actions": [
-        {"type": "service_to_service"}
-      ]
-    }
-    )EOF";
-
-  const std::string request_headers_json = R"EOF(
-    {
-      "domain": "foobar",
-      "actions": [
-        {
-          "type": "request_headers",
-          "header_name": "x-header-name",
-          "descriptor_key" : "my_header_name"
-        }
-      ]
-    }
-    )EOF";
-
-  const std::string address_json = R"EOF(
-    {
-      "domain": "foo",
-      "actions": [
-        {"type": "remote_address"}
-      ]
     }
     )EOF";
 
@@ -115,10 +58,12 @@ public:
   Buffer::OwnedImpl data_;
   Stats::IsolatedStoreImpl stats_store_;
   NiceMock<Runtime::MockLoader> runtime_;
+  std::vector<std::reference_wrapper<Router::RateLimitPolicyEntry>> rate_limit_policies_;
+  Router::TestRateLimitPolicyEntry rate_limit_policy_entry_;
 };
 
 TEST_F(HttpRateLimitFilterTest, NoRoute) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
 
   EXPECT_CALL(filter_callbacks_.route_table_, routeForRequest(_)).WillOnce(Return(nullptr));
 
@@ -128,15 +73,17 @@ TEST_F(HttpRateLimitFilterTest, NoRoute) {
 }
 
 TEST_F(HttpRateLimitFilterTest, NoLimiting) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
 
+  EXPECT_CALL(filter_callbacks_.route_table_.route_entry_.rate_limit_policy_,
+              getApplicableRateLimit_(0)).WillOnce(testing::ReturnRef(rate_limit_policies_));
   EXPECT_EQ(FilterHeadersStatus::Continue, filter_->decodeHeaders(request_headers_, false));
   EXPECT_EQ(FilterDataStatus::Continue, filter_->decodeData(data_, false));
   EXPECT_EQ(FilterTrailersStatus::Continue, filter_->decodeTrailers(request_headers_));
 }
 
 TEST_F(HttpRateLimitFilterTest, RuntimeDisabled) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
 
   EXPECT_CALL(runtime_.snapshot_, featureEnabled("ratelimit.http_filter_enabled", 100))
       .WillOnce(Return(false));
@@ -146,10 +93,19 @@ TEST_F(HttpRateLimitFilterTest, RuntimeDisabled) {
 }
 
 TEST_F(HttpRateLimitFilterTest, OkResponse) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
   InSequence s;
 
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
+
+  std::vector<::RateLimit::Descriptor> descriptor{
+      {{{"to_cluster", "fake_cluster"}}},
+      {{{"to_cluster", "fake_cluster"}, {"from_cluster", "service_cluster"}}}};
+  EXPECT_CALL(filter_callbacks_.route_table_.route_entry_.rate_limit_policy_,
+              getApplicableRateLimit_(0)).WillOnce(testing::ReturnRef(rate_limit_policies_));
+
+  EXPECT_CALL(rate_limit_policy_entry_, populateDescriptors_(_, _, _, _, _))
+      .WillOnce(SetArgReferee<1>(descriptor));
 
   EXPECT_CALL(*client_,
               limit(_, "foo",
@@ -172,7 +128,7 @@ TEST_F(HttpRateLimitFilterTest, OkResponse) {
 }
 
 TEST_F(HttpRateLimitFilterTest, ImmediateOkResponse) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
   InSequence s;
 
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
@@ -196,11 +152,13 @@ TEST_F(HttpRateLimitFilterTest, ImmediateOkResponse) {
 }
 
 TEST_F(HttpRateLimitFilterTest, ErrorResponse) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
   InSequence s;
 
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
 
+  // EXPECT_CALL(filter_callbacks_.route_table_.route_entry_.rate_limit_policy_,
+  // getApplicableRateLimit_(0)).WillOnce(testing::ReturnRef(empty_vector));
   EXPECT_CALL(*client_, limit(_, _, _, _))
       .WillOnce(WithArgs<0>(Invoke([&](::RateLimit::RequestCallbacks& callbacks)
                                        -> void { request_callbacks_ = &callbacks; })));
@@ -217,7 +175,7 @@ TEST_F(HttpRateLimitFilterTest, ErrorResponse) {
 }
 
 TEST_F(HttpRateLimitFilterTest, LimitResponse) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
   InSequence s;
 
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
@@ -239,7 +197,7 @@ TEST_F(HttpRateLimitFilterTest, LimitResponse) {
 }
 
 TEST_F(HttpRateLimitFilterTest, LimitResponseRuntimeDisabled) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
   InSequence s;
 
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
@@ -264,7 +222,7 @@ TEST_F(HttpRateLimitFilterTest, LimitResponseRuntimeDisabled) {
 }
 
 TEST_F(HttpRateLimitFilterTest, ResetDuringCall) {
-  SetUpTest(service_to_service_json);
+  SetUpTest(filter_config);
   InSequence s;
 
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
@@ -280,7 +238,7 @@ TEST_F(HttpRateLimitFilterTest, ResetDuringCall) {
 }
 
 TEST_F(HttpRateLimitFilterTest, RequestHeaderOkResponse) {
-  SetUpTest(request_headers_json);
+  SetUpTest(filter_config);
 
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
 
@@ -303,7 +261,7 @@ TEST_F(HttpRateLimitFilterTest, RequestHeaderOkResponse) {
 }
 
 TEST_F(HttpRateLimitFilterTest, RateLimitKeyOkResponse) {
-  SetUpTest(request_headers_json);
+  SetUpTest(filter_config);
 
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.route_key_ = "test_key";
@@ -328,7 +286,7 @@ TEST_F(HttpRateLimitFilterTest, RateLimitKeyOkResponse) {
 }
 
 TEST_F(HttpRateLimitFilterTest, NoRateLimitHeaderMatch) {
-  SetUpTest(request_headers_json);
+  SetUpTest(filter_config);
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
 
   EXPECT_CALL(*client_, limit(_, _, _, _)).Times(0);
@@ -338,30 +296,8 @@ TEST_F(HttpRateLimitFilterTest, NoRateLimitHeaderMatch) {
   EXPECT_EQ(FilterTrailersStatus::Continue, filter_->decodeTrailers(request_headers_));
 }
 
-TEST_F(HttpRateLimitFilterTest, AddressRateLimiting) {
-  SetUpTest(address_json);
-  filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
-
-  std::string address = "10.0.0.1";
-  EXPECT_CALL(filter_callbacks_, downstreamAddress()).WillOnce(ReturnRef(address));
-  EXPECT_CALL(*client_, limit(_, "foo", testing::ContainerEq(std::vector<::RateLimit::Descriptor>{
-                                            {{{"remote_address", address}}}}),
-                              ""))
-      .WillOnce(WithArgs<0>(Invoke([&](::RateLimit::RequestCallbacks& callbacks)
-                                       -> void { request_callbacks_ = &callbacks; })));
-
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->decodeHeaders(request_headers_, false));
-  EXPECT_EQ(FilterDataStatus::StopIterationAndBuffer, filter_->decodeData(data_, false));
-  EXPECT_EQ(FilterTrailersStatus::StopIteration, filter_->decodeTrailers(request_headers_));
-
-  EXPECT_CALL(filter_callbacks_, continueDecoding());
-  request_callbacks_->complete(::RateLimit::LimitStatus::OK);
-
-  EXPECT_EQ(1U, stats_store_.counter("cluster.fake_cluster.ratelimit.ok").value());
-}
-
 TEST_F(HttpRateLimitFilterTest, RouteAddressRateLimiting) {
-  SetUpTest(address_json);
+  SetUpTest(filter_config);
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.route_key_ = "test_key";
 
@@ -386,7 +322,7 @@ TEST_F(HttpRateLimitFilterTest, RouteAddressRateLimiting) {
 }
 
 TEST_F(HttpRateLimitFilterTest, NoAddressRateLimiting) {
-  SetUpTest(address_json);
+  SetUpTest(filter_config);
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
 
   EXPECT_CALL(filter_callbacks_, downstreamAddress()).WillOnce(ReturnRef(EMPTY_STRING));
@@ -399,9 +335,8 @@ TEST_F(HttpRateLimitFilterTest, NoAddressRateLimiting) {
 }
 
 TEST_F(HttpRateLimitFilterTest, RateLimitDisabledForRouteKey) {
-  SetUpTest(request_headers_json);
+  SetUpTest(filter_config);
 
-  filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.do_global_limiting_ = true;
   filter_callbacks_.route_table_.route_entry_.rate_limit_policy_.route_key_ = "test_key";
   ON_CALL(runtime_.snapshot_, featureEnabled("ratelimit.test_key.http_filter_enabled", 100))
       .WillByDefault(Return(false));
