+++ test/common/router/config_impl_test.cc
@@ -5,7 +5,6 @@
 
 #include "test/mocks/http/mocks.h"
 #include "test/mocks/upstream/mocks.h"
-#include "test/mocks/ratelimit/mocks.h"
 #include "test/mocks/router/mocks.h"
 #include "test/mocks/runtime/mocks.h"
 #include "test/test_common/utility.h"
@@ -924,303 +923,6 @@ TEST(NullConfigImplTest, All) {
   EXPECT_EQ(0UL, config.responseHeadersToAdd().size());
   EXPECT_EQ(0UL, config.responseHeadersToRemove().size());
   EXPECT_FALSE(config.usesRuntime());
-}
-
-TEST(RateLimitConfiguration, MissingActions) {
-  std::string json = R"EOF(
-{
-"virtual_hosts": [
-{
-  "name": "www2",
-  "domains": ["www.lyft.com"],
-  "routes": [
-    {
-      "prefix": "/",
-      "cluster": "www2",
-      "rate_limit": {
-        "global": true
-      },
-      "rate_limits": [
-      { }
-      ]
-    }
-  ]
-}
-]
-}
-)EOF";
-
-  Json::ObjectPtr loader = Json::Factory::LoadFromString(json);
-  NiceMock<Runtime::MockLoader> runtime;
-  NiceMock<Upstream::MockClusterManager> cm;
-  EXPECT_THROW(ConfigImpl(*loader, runtime, cm), EnvoyException);
-}
-
-TEST(RateLimitConfiguration, BadType) {
-  std::string json = R"EOF(
-{
-"virtual_hosts": [
-  {
-    "name": "www2",
-    "domains": ["www.lyft.com"],
-    "routes": [
-      {
-        "prefix": "/",
-        "cluster": "www2",
-        "rate_limits": [
-        { "actions":[ {"type": "bad_type"}] }
-        ]
-      }
-    ]
-  }
-]
-}
-)EOF";
-
-  Json::ObjectPtr loader = Json::Factory::LoadFromString(json);
-  NiceMock<Runtime::MockLoader> runtime;
-  NiceMock<Upstream::MockClusterManager> cm;
-  EXPECT_THROW(ConfigImpl(*loader, runtime, cm), EnvoyException);
-}
-
-TEST(RateLimitConfiguration, ActionsMissingRequiredFields) {
-  std::string json_one = R"EOF(
-{
-"virtual_hosts": [
-  {
-    "name": "www2",
-    "domains": ["www.lyft.com"],
-    "routes": [
-      {
-        "prefix": "/",
-        "cluster": "www2",
-        "rate_limits": [
-        { "actions":[ {"type": "request_headers"}] }
-        ]
-      }
-    ]
-  }
-]
-}
-)EOF";
-
-  Json::ObjectPtr loader = Json::Factory::LoadFromString(json_one);
-  NiceMock<Runtime::MockLoader> runtime;
-  NiceMock<Upstream::MockClusterManager> cm;
-  EXPECT_THROW(ConfigImpl(*loader, runtime, cm), EnvoyException);
-
-  std::string json_two = R"EOF(
-{
-"virtual_hosts": [
-  {
-    "name": "www2",
-    "domains": ["www.lyft.com"],
-    "routes": [
-      {
-        "prefix": "/",
-        "cluster": "www2",
-        "rate_limits": [
-        { "actions":[
-            {"type": "request_headers",
-              "header_name" : "test"}]
-          }
-        ]
-      }
-    ]
-  }
-]
-}
-)EOF";
-  loader = Json::Factory::LoadFromString(json_two);
-  EXPECT_THROW(ConfigImpl(*loader, runtime, cm), EnvoyException);
-  std::string json_three = R"EOF(
-{
-"virtual_hosts": [
-{
-  "name": "www2",
-  "domains": ["www.lyft.com"],
-  "routes": [
-    {
-      "prefix": "/",
-      "cluster": "www2",
-      "rate_limits": [
-      { "actions":[
-          {"type": "request_headers",
-            "descriptor_key" : "test" }]
-        }
-      ]
-    }
-  ]
-}
-]
-}
-)EOF";
-  loader = Json::Factory::LoadFromString(json_three);
-  EXPECT_THROW(ConfigImpl(*loader, runtime, cm), EnvoyException);
-}
-
-TEST(RateLimitConfiguration, RemoteAddress) {
-  std::string json = R"EOF(
-{
-"virtual_hosts": [
-  {
-    "name": "www2",
-    "domains": ["www.lyft.com"],
-    "routes": [
-      {
-        "prefix": "/foo",
-        "cluster": "www2",
-        "rate_limit": {
-          "global": true
-        },
-        "rate_limits": [
-        { "actions":[ {"type": "remote_address"}] }
-        ]
-      },
-      {
-        "prefix": "/bar",
-        "cluster": "www2"
-      }
-    ]
-  }
-]
-}
-)EOF";
-
-  Json::ObjectPtr loader = Json::Factory::LoadFromString(json);
-  NiceMock<Runtime::MockLoader> runtime;
-  NiceMock<Upstream::MockClusterManager> cm;
-  ConfigImpl config(*loader, runtime, cm);
-  NiceMock<Http::MockStreamDecoderFilterCallbacks> callbacks;
-  NiceMock<Router::MockRouteEntry> route;
-  NiceMock<RateLimit::MockFilterConfig> filter_config;
-  Http::TestHeaderMapImpl header;
-  route.rate_limit_policy_.route_key_ = "";
-  std::string address = "10.0.0.1";
-  EXPECT_CALL(callbacks, downstreamAddress()).WillOnce(ReturnRef(address));
-
-  EXPECT_EQ(0u, config.routeForRequest(genHeaders("www.lyft.com", "/bar", "GET"), 0)
-                    ->rateLimitPolicy()
-                    .getApplicableRateLimit()
-                    .size());
-  std::vector<std::reference_wrapper<RateLimitPolicyEntry>> rate_limits =
-      config.routeForRequest(genHeaders("www.lyft.com", "/foo", "GET"), 0)
-          ->rateLimitPolicy()
-          .getApplicableRateLimit();
-  EXPECT_EQ(1u, rate_limits.size());
-  std::vector<::RateLimit::Descriptor> descriptors;
-  for (const RateLimitPolicyEntry& rate_limit : rate_limits) {
-    rate_limit.populateDescriptors(route, descriptors, filter_config, header, callbacks);
-  }
-  std::vector<::RateLimit::Descriptor> expected_descriptors = {{{{"remote_address", address}}}};
-  EXPECT_EQ(descriptors, expected_descriptors);
-}
-
-TEST(RateLimitConfiguration, ServiceToService) {
-  std::string json = R"EOF(
-{
-"virtual_hosts": [
-  {
-    "name": "www2",
-    "domains": ["www.lyft.com"],
-    "routes": [
-      {
-        "prefix": "/foo",
-        "cluster": "www2",
-        "rate_limit": {
-          "global": true
-        },
-        "rate_limits": [
-        { "actions":[ {"type": "service_to_service"}] }
-        ]
-      },
-      {
-        "prefix": "/bar",
-        "cluster": "www2"
-      }
-    ]
-  }
-]
-}
-)EOF";
-
-  Json::ObjectPtr loader = Json::Factory::LoadFromString(json);
-  NiceMock<Runtime::MockLoader> runtime;
-  NiceMock<Upstream::MockClusterManager> cm;
-  ConfigImpl config(*loader, runtime, cm);
-  NiceMock<Http::MockStreamDecoderFilterCallbacks> callbacks;
-  NiceMock<Router::MockRouteEntry> route;
-  NiceMock<RateLimit::MockFilterConfig> filter_config;
-  Http::TestHeaderMapImpl header;
-  route.rate_limit_policy_.route_key_ = "my_route";
-  filter_config.local_service_cluster_ = "service_cluster";
-  std::vector<std::reference_wrapper<RateLimitPolicyEntry>> rate_limits =
-      config.routeForRequest(genHeaders("www.lyft.com", "/foo", "GET"), 0)
-          ->rateLimitPolicy()
-          .getApplicableRateLimit();
-  EXPECT_EQ(1u, rate_limits.size());
-  std::vector<::RateLimit::Descriptor> descriptors;
-  for (const RateLimitPolicyEntry& rate_limit : rate_limits) {
-    rate_limit.populateDescriptors(route, descriptors, filter_config, header, callbacks);
-  }
-  std::vector<::RateLimit::Descriptor> expected_descriptors = {
-      {{{"to_cluster", "fake_cluster"}}},
-      {{{"to_cluster", "fake_cluster"}, {"from_cluster", "service_cluster"}}}};
-  EXPECT_EQ(descriptors, expected_descriptors);
-}
-TEST(RateLimitConfiguration, RequestHeaders) {
-  std::string json = R"EOF(
-{
-"virtual_hosts": [
-  {
-    "name": "www2",
-    "domains": ["www.lyft.com"],
-    "routes": [
-      {
-        "prefix": "/",
-        "cluster": "www2",
-        "rate_limit": {
-          "global": true
-        },
-        "rate_limits": [
-        { "actions": [
-          {
-            "type": "request_headers",
-            "header_name": "x-header-name",
-            "descriptor_key" : "my_header_name"
-          }
-        ]}
-        ]
-      }
-    ]
-  }
-]
-}
-)EOF";
-
-  Json::ObjectPtr loader = Json::Factory::LoadFromString(json);
-  NiceMock<Runtime::MockLoader> runtime;
-  NiceMock<Upstream::MockClusterManager> cm;
-  ConfigImpl config(*loader, runtime, cm);
-  NiceMock<Http::MockStreamDecoderFilterCallbacks> callbacks;
-  NiceMock<Router::MockRouteEntry> route;
-  NiceMock<RateLimit::MockFilterConfig> filter_config;
-  Http::TestHeaderMapImpl header{{"x-header-name", "test_value"}};
-  route.rate_limit_policy_.route_key_ = "test_key";
-  filter_config.local_service_cluster_ = "service_cluster";
-  std::vector<std::reference_wrapper<RateLimitPolicyEntry>> rate_limits =
-      config.routeForRequest(genHeaders("www.lyft.com", "/foo", "GET"), 0)
-          ->rateLimitPolicy()
-          .getApplicableRateLimit();
-  EXPECT_EQ(1u, rate_limits.size());
-  std::vector<::RateLimit::Descriptor> descriptors;
-  for (const RateLimitPolicyEntry& rate_limit : rate_limits) {
-    rate_limit.populateDescriptors(route, descriptors, filter_config, header, callbacks);
-  }
-  std::vector<::RateLimit::Descriptor> expected_descriptors = {
-      {{{"my_header_name", "test_value"}}},
-      {{{"route_key", "test_key"}, {"my_header_name", "test_value"}}}};
-  EXPECT_EQ(descriptors, expected_descriptors);
 }
 
 } // Router
